create and finalized scripts for face recognizion in python with opencv. changed hide interval of webui

+++ stuff/scripts/face-detect.py
@@ -2,12 +2,12 @@
 
 import sys
 from time import sleep, gmtime, strftime
-import cv2
-import picamera
-import os
-
-import threading, logging, time, ConfigParser
+import cv2, picamera, os, threading, logging, time, ConfigParser
+import numpy as np
+from PIL import Image
 from kafka import KafkaProducer
+import scipy.ndimage
+
 
 Config = ConfigParser.ConfigParser()
 Config.read('./config.ini')
@@ -16,38 +16,96 @@ kafkaHost = Config.get('Kafka', 'host')
 kafkaPort = Config.get('Kafka', 'port')
 topic = Config.get('face-producer', 'topic')
 
+recognizerFile= '/home/pi/facedetect/generated.rec'
 
 
-def find_a_face(image):
-    face_cascade = cv2.CascadeClassifier('/home/pi/opencv-3.0.0/data/haarcascades/haarcascade_frontalface_default.xml')
+cascade_path = '/home/pi/opencv-3.0.0/data/haarcascades/haarcascade_frontalface_default.xml'
+eyes_path= '/home/pi/opencv-3.0.0/data/haarcascades/haarcascade_eye.xml'
+noses_path= '/home/pi/opencv-3.0.0/data/haarcascades/Nariz.xml'
 
-    gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
+face_cascade = cv2.CascadeClassifier(cascade_path)
+eye_cascade = cv2.CascadeClassifier(eyes_path)
+nose_cascade = cv2.CascadeClassifier(noses_path)
 
-    faces = face_cascade.detectMultiScale(
-        gray,
-        scaleFactor=1.1,
-        minNeighbors=5,
-        minSize=(30, 30),
-        flags = cv2.CASCADE_SCALE_IMAGE
-    )
+# For face recognition we will the the LBPH Face Recognizer
+recognizer = cv2.face.createLBPHFaceRecognizer()
 
-    # iterate over all identified faces and try to find eyes
-    for (x, y, w, h) in faces:
+show_windows=False
+
+
+prediction_lower_limit = 80.
+prediction_upper_limit = 100.
+
+def really_a_face (inner_face):
+
+    eyes = eye_cascade.detectMultiScale(inner_face, minSize=(10, 10))
+    if len(eyes) >= 2 :
+
+        if show_windows:
+
+            for (ex,ey,ew,eh) in eyes[:2]:
+                cv2.rectangle(inner_face,(ex,ey),(ex+ew,ey+eh),(255,0,0),2)
+
+            cv2.imshow("Adding faces to traning set...", inner_face)
+            cv2.waitKey(5)
+
         return True
+
+    return False
+
+
+
+
+def check_image(new_image):
+    nbrs = []
+    predict_image_pil = Image.open(new_image).convert('L')
+    predict_image = np.array(predict_image_pil, 'uint8')
+    faces = face_cascade.detectMultiScale(predict_image,
+                                         scaleFactor=1.1,
+                                         minNeighbors=5,
+                                         minSize=(100, 100),
+                                         flags = cv2.CASCADE_SCALE_IMAGE)
+    for (x, y, w, h) in faces:
+        inner_face= predict_image[y: y + h, x: x + w]
+
+        if really_a_face(inner_face):
+            resized= scipy.misc.imresize(inner_face, ( 100. / h ))
+            nbr_predicted, conf = recognizer.predict(resized)
+            print "nbr: {}, conf: {}".format(nbr_predicted, conf)
+
+            if prediction_lower_limit <= conf <= prediction_upper_limit:
+                nbrs.append(nbr_predicted)
+
+    return nbrs
+
+
+
+
 
 
 if __name__ == "__main__":
 
-    cam = picamera.PiCamera()
-    producer = KafkaProducer(bootstrap_servers=kafkaHost+':'+kafkaPort)
+    if len(sys.argv) >= 2:
+        show_windows=True
 
-    while True:
-        image_file = '/tmp/face-detect.jpg'
-        cam.capture(image_file)
-        image = cv2.imread(image_file)
-        if find_a_face(image):
-            print( strftime("%Y-%m-%d %H:%M:%S", gmtime()) + ': found a face!')
-            producer.send(topic, b'{"found" : true, "person": "any"}')
+    if os.path.exists(recognizerFile):
+        recognizer.load(recognizerFile)
 
-        os.remove(image_file)
-        sleep(0.1)
+        cam = picamera.PiCamera()
+        producer = KafkaProducer(bootstrap_servers=kafkaHost+':'+kafkaPort)
+
+        while True:
+            image_file = '/tmp/face-detect.jpg'
+            cam.capture(image_file)
+
+            ids= check_image(image_file)
+
+            if ids :
+                print( strftime("%Y-%m-%d %H:%M:%S", gmtime()) + ': found a face! {}'.format(ids))
+                producer.send(topic, b'{"found" : true, "person": "any"}')
+
+            os.remove(image_file)
+            sleep(0.1)
+
+        else:
+            print "no recognizer found."
